This branch brings the ability to add labels/titles to entries. On the command line this isn't much of a necessity but could be a nice feature in a GUI, where the title can mask a big ugly URL.
You can add a title to a new entry with the `--title:` flag, like so:
    $ star -n tag1 tag2 tag3 http://big.ugly.url --title:"This is the title"

The `--title:` flag can appear anywhere in the argument list but the format is rigid. Of course quotes are needed for a title that includes more than one word.

This required a small change to the file format. For each entry the title is stored in the same section as the value, but in the second position since it is optional. The value and title are separated by the same separator as are individual tags and the metadata components.

When filtering, titles are searched and match in the same way the values are.

Titles are included in the output lists between the numbered line and the tags:
    1) http://big.ugly.url
       Title: this is the title
       Tags: tag1, tag2, tag3

And, when editing, you can add, edit, and remove titles from an entry in the same way you can the other components.

* Labels [6/6]
  - [X] Read labels from args
  - [X] Save labels to file
  - [X] Search labels
  - [X] Display labels in list
  - [X] Display labels in editing file
  - [X] Update labels from editing file